The code has been moved from PlayRooms.jsx to Utils index.js
